requirements.txt
flask
tensorflow
rembg
pillow
gunicorn
